chore: :building_construction: Rename Accordion folder
